+++ packages/pangea-multipass/tests/integration/test_github.py
@@ -1,35 +1,35 @@
 import os
 import unittest
 
-from pangea_multipass import GitLabProcessor, GitLabReader, get_document_metadata
+from pangea_multipass import GitHubProcessor, GitHubReader, get_document_metadata
 
-token = os.getenv("GITLAB_ADMIN_TOKEN") or ""
-username = os.getenv("GITLAB_USERNAME") or ""
+token = os.getenv("GITHUB_ADMIN_TOKEN") or ""
+username = os.getenv("GITHUB_USERNAME") or ""
 
 _TOTAL_FILES = 8
 _AUTHORIZED_FILES = 5
 _AUTHORIZED_PROJECTS = 2
 
 
-class TestGitLab(unittest.TestCase):
+class TestGitHub(unittest.TestCase):
     def setUp(self) -> None:
         assert token
         assert username
 
-    def test_gitlab(self) -> None:
-        reader = GitLabReader(token=token)
+    def test_github(self) -> None:
+        reader = GitHubReader(token=token)
         files = reader.load_data()
         assert len(files) == _TOTAL_FILES
 
-        processor = GitLabProcessor(admin_token=token, username=username, get_node_metadata=get_document_metadata)
+        processor = GitHubProcessor(token=token, username=username, get_node_metadata=get_document_metadata)
         filter = processor.get_filter()
         assert len(filter.value) == _AUTHORIZED_PROJECTS
 
         authorized_files = processor.filter(files)
         assert len(authorized_files) == _AUTHORIZED_FILES
 
-    def test_gitlab_pagination(self) -> None:
-        reader = GitLabReader(token=token)
+    def test_github_pagination(self) -> None:
+        reader = GitHubReader(token=token)
 
         repos = reader.get_repos()
         assert len(repos) == 3